+++ frontend/src/features/auth/components/modals/LoginForm.js
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { useCookies } from 'react-cookie';
 import { useNavigate } from 'react-router-dom';
  
-import FindpasswordForm from './findpassword.js';
-import SignUpinfoForm from './Signupinfo.js';
+import FindpasswordForm from './FindPassword.js';
+import SignUpinfoForm from './Signup.js';
 import MyModal from '../../../../shared/components/BaseModal.js';
 import './modules.css';
 import { signInRequest } from '../../../../entities/api/UserApi.js';
